File Generation, etc.
If Locker() can’t read either data.csv or hash.txt it will call
initFile() which checks of those files exists, and if not generates
them in the proper format.

After first successful login, a new entry is added to the password list
containing the master password and logged in user for username.